vendor_management/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
flake8==6.1.0
python-dotenv==1.0.0
black==23.11.0
psycopg2-binary==2.9.9
django-shortuuidfield==0.1.3
drf-spectacular==0.27.0
coverage==7.3.3
